nancy_iBiz 发布系统代码 [后台服务体系，iBiz软件生产管理]

+++ app_Web/src/widgets/ibizpro-index/index-mob-list/index-mob-list-model.ts
@@ -41,7 +41,7 @@ export default class IndexMobModel {
 			{
 				name: 'srfkey',
 				prop: 'indexid',
-				dataType: 'BIGINT',
+				dataType: 'INT',
 			},
 			{
 				name: 'srfmajortext',
