+++ app_Web/src/widgets/ibizpro-index/index-type-dataview/index-type-dataview-model.ts
@@ -22,7 +22,7 @@ export default class IndexTypeModel {
 			{
 				name: 'srfkey',
 				prop: 'indexid',
-				dataType: 'BIGINT',
+				dataType: 'INT',
 			},
 
 			{
